Up from Linux 3.10.63 to Linux 3.10.69
This commits carries Updates  3.10.64, 3.10.65, 3.10.66, 3.10.67, 3.10.68 and 3.10.69.
Aside various reports that are not functional for the S5plus And patches from the android community.

Special thanks to @placiano and @Pafcholini for patches and reverts in emotion kernel for Note4. (cherry-pic from 523956fab74ab89358aa2502929c9546fc396c94 to 2a4c7c4d21eaf65546da2494553e4741f827fb99 in the branch dev)

The migration of 63 to 69 carries many complications for this terminal.

Signed-off-by: Javier Sayago <admin@lonasdigital.com>