workqueue: use 'pool->cpu < 0' to stand for an unbound pool
There is a piece of sanity checks code in the put_unbound_pool().
The meaning of this code is 'if it is not an unbound pool, it will complain
and return' IIUC. But the code uses 'pool->flags & POOL_DISASSOCIATED'
imprecisely due to a non-unbound pool may also have this flags.

We should use 'pool->cpu < 0' to stand for an unbound pool, so we covert the
code to it.

There is no strictly wrong if we still keep 'pool->flags & POOL_DISASSOCIATED'
here, but it is just a noise if we keep it:
  1) we focus on 'unbound' here, not '[dis]association'.
  2) 'pool->cpu < 0' already implies 'pool->flags & POOL_DISASSOCIATED'.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>
Signed-off-by: Tejun Heo <tj@kernel.org>